+++ src/astrameter/config/config_loader_test.py
@@ -606,6 +606,7 @@ def test_create_tibber_pulse_powermeter() -> None:
     assert pm.node_id == "1"
     assert pm.user == "admin"
     assert pm.timeout == 5.0
+    assert pm.force_polling is False
 
     config["TIBBER_PULSE_2"] = {
         "IP": "127.0.0.1",
@@ -613,6 +614,7 @@ def test_create_tibber_pulse_powermeter() -> None:
         "NODE_ID": "2",
         "USER": "root",
         "TIMEOUT": "10",
+        "FORCE_POLLING": "true",
         "OBIS_POWER_CURRENT": "0100100700ff",
     }
     pm = create_tibber_pulse_powermeter("TIBBER_PULSE_2", config)
@@ -620,6 +622,7 @@ def test_create_tibber_pulse_powermeter() -> None:
     assert pm.node_id == "2"
     assert pm.user == "root"
     assert pm.timeout == 10.0
+    assert pm.force_polling is True
     assert pm._obis_current == "0100100700ff"